Added file system with documentation at lines 31-39, for example you can now use /projects/Boxy/workplan instead of /projects/model.stpnc/workplan. Also created some temporary projects endpoints that will need to be updated.
