+++ src/server/api/v2/temp_doc.md
@@ -26,3 +26,13 @@ The endpoint structure of these files
 			+ {wsID}
 
 In order to expand these endpoints, the typical notation is property/id/.
+
+
+Must setup a data folder in the root directory that contains a pathmap.json file with example structure:
+{
+	"ncId": "Path to .stpnc file within data directory e.g. "C:/Users/Nick/Documents/STEP Tools/StepNCViewer/data/Boxy/model.stpnc" "
+}
+In order to get the file path for a specific project use:
+
+var file = require('./file'); //This is if the file.js is in the same directory as your .js file
+file.getPath(ncId) //This will give you the absolute path for the given ncId based on the pathmap.json
